Remove client_address from can_handle method and add TftpError.
The considerations here are the same as for the HttpRequestHandler:
A handler should not use the client address to decide whether it can
handle a request.

If it later finds out that it does not want to handle the request (e.g.
because the client lacks the necessary permissions) it can still signal
this problem by raising a TftpError.