Disable template cache when files should not be cached at all.
This commit adds a reminder to the documentation of the data source
that disabling the data source cache does not disable the template
cache.

It also disables the template cache for one specific unit test that
would otherwise fail when running on a file system where the time stamp
is not very precise.