Invalidate cache when imported template changes.
There was a bug in the caching of the YAML target data-source. If a
template file imported another template file and that imported file
changed, this change was not detected and cached data was used
instead of recompiling the data.

Unfortunately, there was no simple fix for this: There is no way to
get the list of imported (or included) templates for a Jinja template.
Even if there was a way, there would be no guarantee that a similar
situation could be detected for all possible template engines.

For this reason, we changed the caching architecture: Now, we always
render the templates. With Jinja, this is not that expensive because
compiled templates are cached and the logic of templates themselves
typically is not very expensive. We still use a cache in order to
avoid the YAML parse and the merge process of the different files when
there are no changes.