Improve file version calculation.
Now, we also include the device number, the inode number or file index,
and the file size. This makes it a bit more likely that we detect
certain kinds of changes that happen too quickly to detect them because
of a change of the time stamp.

We also now use the time stamp in nanoseconds instead of using the
floating point value, though there is probably no practical advantage to
that.