Normalize IP addresses before comparison.
Before, we would only apply IPv6 to IPv4 conversion to the actual client
IP address when necessary and not normalize the expected client
address(es) at all.

Now, we use the vinegar.transform.ip_address.normalize function on both
the actual and the expected address(es). This simplifies our code and
makes the comparison more reliable.